Fix cbor parsing bug
Previously, when options were parsed, the corresponding values were not
being consumed if they were not one of the recognized strings.
Authenticators may provide additional options, in which case the value
should be consumed and ignored. Otherwise, an error will be thrown
because the decoder encounters a boolean instead of the next integer key
of the map.

This small patch fixes this bug.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>